fix: cli should enforce args
