+++ .projen/tasks.json
@@ -126,7 +126,7 @@
       "name": "integ:exec:lambda:fail",
       "steps": [
         {
-          "exec": "ts-node -P tsconfig.dev.json src/cli.ts --app test/exec.integ.snapshot exec integ-cdk-exec/Function"
+          "exec": "ts-node -P tsconfig.dev.json src/cli/cdk-exec.ts --app test/exec.integ.snapshot exec integ-cdk-exec/Function"
         }
       ]
     },
@@ -134,7 +134,7 @@
       "name": "integ:exec:lambda:succeed",
       "steps": [
         {
-          "exec": "ts-node -P tsconfig.dev.json src/cli.ts --app test/exec.integ.snapshot exec integ-cdk-exec/Function --input '{\"succeed\":true}'"
+          "exec": "ts-node -P tsconfig.dev.json src/cli/cdk-exec.ts --app test/exec.integ.snapshot exec integ-cdk-exec/Function --input '{\"succeed\":true}'"
         }
       ]
     },
@@ -142,7 +142,7 @@
       "name": "integ:exec:sfn:fail",
       "steps": [
         {
-          "exec": "ts-node -P tsconfig.dev.json src/cli.ts --app test/exec.integ.snapshot exec integ-cdk-exec/StateMachine"
+          "exec": "ts-node -P tsconfig.dev.json src/cli/cdk-exec.ts --app test/exec.integ.snapshot exec integ-cdk-exec/StateMachine"
         }
       ]
     },
@@ -150,7 +150,7 @@
       "name": "integ:exec:sfn:succeed",
       "steps": [
         {
-          "exec": "ts-node -P tsconfig.dev.json src/cli.ts --app test/exec.integ.snapshot exec integ-cdk-exec/StateMachine --input '{\"succeed\":true}'"
+          "exec": "ts-node -P tsconfig.dev.json src/cli/cdk-exec.ts --app test/exec.integ.snapshot exec integ-cdk-exec/StateMachine --input '{\"succeed\":true}'"
         }
       ]
     },
