+++ test/exec.integ.snapshot/cdk.out
@@ -1,1 +1,1 @@
-{"version":"16.0.0"}
+{"version":"15.0.0"}
