+++ test/exec.integ.snapshot/integ-cdk-exec.assets.json
@@ -1,7 +1,7 @@
 {
-  "version": "16.0.0",
+  "version": "15.0.0",
   "files": {
-    "ab7ed3e4c4b47c50728145e113d924e757cfbf17f1302b74943e892e8199d415": {
+    "4c8b33609a261af561cfcb236b963c62cd5725d000edc30b57c9a3f0d50df71b": {
       "source": {
         "path": "integ-cdk-exec.template.json",
         "packaging": "file"
@@ -9,7 +9,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "ab7ed3e4c4b47c50728145e113d924e757cfbf17f1302b74943e892e8199d415.json",
+          "objectKey": "4c8b33609a261af561cfcb236b963c62cd5725d000edc30b57c9a3f0d50df71b.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
